Multi-threading and long form support
A new thread will be opened to send the Discord embeds. This is to minimize the delay from sending the embeds to Discord. Additionally, long forms are now supported -- instead of sending the whole form in one message, it'll split the form whenever it exceeds 4096 characters. Finally, rather than showing the time via plain text, it will now use Discord's formatting.